+++ src/components/CryptoMarket.vue
@@ -2,7 +2,8 @@
   <div :id="container" ref="tradingview" />
 </template>
 <script lang="ts">
-import { defineComponent, onMounted, ref } from 'vue';
+import { defineComponent } from 'vue';
+import useInitWidget from '../composable/useInitWidget';
 
 export default defineComponent({
   name: 'CryptoMarket',
@@ -21,41 +22,13 @@ export default defineComponent({
     },
   },
   setup(props) {
-    const container = ref('tradingview-crypto-market');
-    const scriptID = ref('tradingview-crypto-market-script');
-    const tradingview = ref<HTMLDivElement>();
-
-    const canUseDOM = () => {
-      return typeof window !== 'undefined' && window.document && window.document.createElement;
-    };
-
-    const getScriptElement = () => {
-      return document.getElementById(scriptID.value);
-    };
-
-    const scriptExists = () => {
-      return getScriptElement() !== null;
-    };
-
-    const appendScript = () => {
-      if (!canUseDOM()) return;
-      if (scriptExists()) return;
-
-      const script = document.createElement('script');
-      script.id = scriptID.value;
-      script.type = 'text/javascript';
-      script.async = true;
-      script.src = 'https://s3.tradingview.com/external-embedding/embed-widget-screener.js';
-      script.textContent = JSON.stringify(props.options);
-      if (tradingview.value) tradingview.value.appendChild(script);
-    };
-
-    onMounted(() => {
-      setTimeout(() => {
-        appendScript();
-      }, 300);
-    });
-    return { canUseDOM, scriptExists, container, tradingview };
+    const { container, tradingview } = useInitWidget(
+      props.options,
+      'tradingview-crypto-market',
+      'tradingview-crypto-market-script',
+      'https://s3.tradingview.com/external-embedding/embed-widget-screener.js'
+    );
+    return { container, tradingview };
   },
 });
 </script>
